removed zip and fixed some warnings

+++ Astar/GridTile.cpp
@@ -17,12 +17,12 @@ GridTile::GridTile(sf::Vector2f t_pos, sf::Vector2f t_size, int t_index, sf::Fon
 	m_tile.setOutlineColor(sf::Color::Black);
 	m_tile.setOutlineThickness(-1.0f);
 
-	int adjustX = (m_pos.x / m_tile.getSize().x) + 1;
-	int adjustY = (m_pos.y / m_tile.getSize().y) + 1;
+	int adjustX = (int)((m_pos.x / m_tile.getSize().x) + 1);
+	int adjustY = (int)((m_pos.y / m_tile.getSize().y) + 1);
 
-	int x = (m_pos.x / (m_tile.getSize().x / 2) - adjustX) * 100;
-	int y = (m_pos.y / (m_tile.getSize().y / 2) - adjustY) * 100;
-	m_pathPos = sf::Vector2f(x, y);
+	int x = (int)((m_pos.x / (m_tile.getSize().x / 2) - adjustX) * 100);
+	int y = (int)((m_pos.y / (m_tile.getSize().y / 2) - adjustY) * 100);
+	m_pathPos = sf::Vector2f((float)x, (float)y);
 
 	m_text.setFont(m_font);
 	m_text.setCharacterSize(12);
